issue #189 test: rename msw handler users/calendar to user/calendar
